requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.4
Django==4.2.5
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Markdown==3.4.4
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
